refactor: WishItemT imageUrl 타입 optional 대신 string | null로 변경
